refactor: defined folder structure
