+++ src/components/Console/Console.style.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const Console = styled.div`
+  width: 548px;
+  height: 703px;
+  padding: 20px;
+  overflow: auto;
+  border-radius: 4px;
+  color: #ffffff;
+  background-color: #2a3949;
+  font-family: monospace;
+  white-space: pre;
+`;
