+++ src/components/Console/Console.tsx
@@ -1,0 +1,3 @@
+import { Console } from "./Console.style";
+
+export default Console;
