+++ src/components/Console/index.ts
@@ -1,0 +1,3 @@
+import Console from "./Console";
+
+export default Console;
