+++ src/components/ImageMapEditor/index.ts
@@ -1,0 +1,3 @@
+import ImageMapEditor from "./ImageMapEditor";
+
+export default ImageMapEditor;
