+++ src/components/index.ts
@@ -1,0 +1,4 @@
+import Console from "./Console";
+import ImageMapEditor from "./ImageMapEditor";
+
+export { Console, ImageMapEditor };
