+++ src/layouts/Layout.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const Layout = styled.div`
+  display: flex;
+  justify-content: center;
+  gap: 135px;
+  flex-wrap: wrap;
+`;
+
+export default Layout;
