+++ src/types/coordinate.ts
@@ -1,0 +1,7 @@
+export interface Coordinate {
+  id: string;
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
